requirements.txt
fastapi
uvicorn
python-dotenv
werkzeug
requests
python-multipart